# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: PESATRIX Open Education Infrastructure API
  description: Reusable, secure, interoperable digital infrastructure for structured
    educational events and information in low-resource environments. No personal data
    is stored by this system.
  version: 0.1.0
paths:
...
  /api/v1/auth/token:
    post:
      tags:
      - auth
      summary: Exchange credentials for access and refresh tokens
      operationId: token_api_v1_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Rotate an access token using a refresh token
      operationId: refresh_token_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/users:
    post:
      tags:
      - users
      summary: Create a user (admin)
      operationId: create_api_v1_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - users
      summary: List users (admin)
      operationId: list_users_api_v1_users_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_UserResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/attendance/rollup:
    post:
      tags:
      - attendance
      summary: Compute (or recompute) attendance records for a school on a date
      operationId: rollup_api_v1_attendance_rollup_post
      parameters:
      - name: school_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: School Id
      - name: day
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Day
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attendance:
    get:
      tags:
      - attendance
      summary: List attendance records
      operationId: list_records_api_v1_attendance_get
      parameters:
      - name: school_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: School Id
      - name: day
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Day
      - name: student_identity_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceRecordResponse'
                title: Response List Records Api V1 Attendance Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attendance/summary:
    get:
      tags:
      - attendance
      summary: Attendance summary counts for a school on a date
      operationId: summary_api_v1_attendance_summary_get
      parameters:
      - name: school_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: School Id
      - name: day
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Day
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AttendanceRecordResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        school_id:
          type: string
          format: uuid
          title: School Id
        student_identity_id:
          type: string
          format: uuid
          title: Student Identity Id
        date:
          type: string
          format: date
          title: Date
        status:
          $ref: '#/components/schemas/AttendanceStatus'
        source_event_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Event Id
        derived_at:
          type: string
          format: date-time
          title: Derived At
      type: object
      required:
      - id
      - school_id
      - student_identity_id
      - date
      - status
      - derived_at
      title: AttendanceRecordResponse
    AttendanceStatus:
      type: string
      enum:
      - present
      - late
      - absent
      - excused
      - unrecorded
      title: AttendanceStatus
    AttendanceSummaryEntry:
      properties:
        date:
          type: string
          title: Date
        present:
          type: integer
          title: Present
        late:
          type: integer
          title: Late
        absent:
          type: integer
          title: Absent
        excused:
          type: integer
          title: Excused
        unrecorded:
          type: integer
          title: Unrecorded
        total:
          type: integer
          title: Total
      type: object
      required:
      - date
      - present
      - late
      - absent
      - excused
      - unrecorded
      - total
      title: AttendanceSummaryEntry
    AttendanceSummaryResponse:
      properties:
        school_id:
          type: string
          format: uuid
          title: School Id
        date:
          type: string
          title: Date
        total_students:
          type: integer
          title: Total Students
        entries:
          items:
            $ref: '#/components/schemas/AttendanceSummaryEntry'
          type: array
          title: Entries
      type: object
      required:
      - school_id
      - date
      - total_students
      - entries
      title: AttendanceSummaryResponse
...
    AuthTokenRequest:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      type: object
      required:
      - username
      - password
      title: AuthTokenRequest
    AuthTokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      title: AuthTokenResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
      type: object
      required:
      - message
      title: MessageResponse
...
    PaginatedResponse_UserResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[UserResponse]
    RecordStatus:
      type: string
      enum:
      - active
      - suspended
      - archived
      title: RecordStatus
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshRequest
...
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: ^[a-zA-Z0-9_.-]+$
          title: Username
        password:
          type: string
          maxLength: 256
          minLength: 8
          title: Password
        display_name:
          type: string
          maxLength: 120
          minLength: 1
          title: Display Name
        role:
          type: string
          title: Role
          default: operator
        organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organization Id
        school_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: School Id
      type: object
      required:
      - username
      - password
      - display_name
      title: UserCreate
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        role:
          type: string
          title: Role
        status:
          $ref: '#/components/schemas/RecordStatus'
        organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organization Id
        school_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: School Id
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
      type: object
      required:
      - id
      - created_at
      - updated_at
      - username
      - display_name
      - role
      - status
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError